Create a helpers/mixin and move trim method into it.
